+++ tests/test_image_seek.c
@@ -28,6 +28,6 @@ int main(int ac, char **av)
     }
     nmd_release_frame(f);
 
-    nmd_free(&s);
+    nmd_freep(&s);
     return 0;
 }