+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- `nmd_free()` renamed to `nmd_freep()`
+
 ## [10.0.0] - 2023-04-05
 
 ### Changed